+++ src/main/java/domain/attendance/AttendanceTime.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.attendance;
 
+import domain.holiday.Holiday;
 import java.time.DayOfWeek;
 import java.time.Duration;
 import java.time.LocalDate;
