+++ src/main/java/domain/attendance/AttendanceTimes.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
