+++ src/test/java/controller/CommandTest.java
@@ -1,4 +1,4 @@
-package domain;
+package controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
