+++ src/test/java/domain/attendance/AttendanceStatusTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 import static org.assertj.core.api.Assertions.assertThat;
