+++ src/test/java/domain/attendance/AttendanceTimeTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
