+++ src/test/java/domain/holiday/HolidayTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.holiday;
 
 import static org.assertj.core.api.Assertions.assertThat;
